[refactor] Refactor files related options
* Simplified.
