--- src/options/index.ts
@@ -1,11 +1,0 @@
-import InfoAddonOptions from './InfoAddonOptions'
-
-export { default as InfoAddonOptions } from './InfoAddonOptions'
-
-export const defaultOptions: InfoAddonOptions = {
-  header: true,
-  source: true,
-  styles: {},
-  summary: '',
-  components: null
-}
